Revert changes on select_header()
It was found that the change on select_header() was not necessary because this function should only be called when the PPM marker is present. Packed packet header is allowed only in main header when PPM is used. 

(See A.7.4 in the Part 1 spec; "The packet headers shall be in only one of three places within the codestream. If the PPM marker segment is present, all
the packet headers shall be found in the main header. In this case, the PPT marker segment and packets distributed in the
bit stream of the tile-parts are disallowed.")

So, the modification of this function can be reverted without any regression.